# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Cinema API"
  description: "This is a Cinema API where you can execute CRUD operations on 
  movies, actors, genres, sessions, customers and auditoriums"
servers:
  - url: http://127.0.0.1:5000
...
paths:
...
  /tickets:
    get:
      tags:
        - "Tickets"
      security:
        - bearerAuth: []
      summary: "Get tickets information"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "session_id"
          in: "query"
          description: "Session ID (only as admin)"
          required: false
          schema:
            type: "string"
        - name: "user_id"
          in: "query"
          description: "User ID (only as admin)"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketsOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
    post:
      security:
        - bearerAuth: []
      tags:
        - "Tickets"
      summary: "Create new ticket"
      description: "This can only be done by the logged in user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify session_id"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '405':
          description: "No more seats on chosen session"
          content:
            application/json:
              example:
                message: "No more seats available"
  /tickets/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - "Tickets"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of ticket"
          required: true
          schema:
            type: "integer"
      summary: "Get ticket by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Ticket not found"
          content:
            application/json:
              example:
                message: "Ticket not found"
        '405':
          description: "Admin access required"
          content:
            application/json:
              example:
                message: "Not allowed"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Tickets"
      summary: "Update ticket's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of ticket"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketPatchIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong ticket's id"
          content:
            application/json:
              example:
                message: "Ticket not found"
        '405':
          description: "No more seats on chosen session"
          content:
            application/json:
              example:
                message: "No more seats available"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Tickets"
      summary: "Delete ticket"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the ticket"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong ticket's id"
          content:
            application/json:
              example:
                message: "Ticket not found"
        '405':
          description: "Admin access required"
          content:
            application/json:
              example:
                message: "Not allowed"
  /tickets/inactive:
    get:
      tags:
        - "Tickets"
      security:
        - bearerAuth: [ ]
      summary: "Get current user inactive tickets as user or all inactive tickets as admin"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "session_id"
          in: "query"
          description: "Session ID (only as admin)"
          required: false
          schema:
            type: "string"
        - name: "user_id"
          in: "query"
          description: "User ID (only as admin)"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketsOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TicketOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        session_id:
          type: "integer"
          example: 2
        user_id:
          type: "integer"
          example: 2
        is_active:
          type: "boolean"
          example: true
    TicketsOut:
      type: "array"
      items:
        $ref: '#/components/schemas/TicketOut'
    TicketIn:
      type: "object"
      properties:
        session_id:
          type: "integer"
          example: 2
        user_id:
          type: "integer"
          example: 2
    TicketPatchIn:
      type: "object"
      properties:
        session_id:
          type: "integer"
          example: 2
        user_id:
          type: "integer"
          example: 2
        is_active:
          type: "boolean"
          example: true
...